requirements.txt
pydantic>=2.8
pydantic-ai>=0.3
pypdf>=5.0
python-docx>=1.1
